Move native function declarations into a new file
